requirements.txt
streamlit
pandas
xlsxwriter
pytesseract
Pillow
PyMuPDF

